// src/types/form.ts
export type FormValues = {
  car_id: number;
  name: string;
  phone: string;
  address: string;
  city: string;
  pickUpLocation: string;
  dropOffLocation: string;
  pickUpDate: string;
  dropOffDate: string;
  pickUpTime: string;
  dropOffTime: string;
  cardNumber: string;
  expiryDate: string;
  cardHolder: string;
  cvc: string;
  termsAccepted: boolean;
};